wireshark_extcap/requirements.txt
pyserial==3.5
pyinstaller==6.12.0